+++ vitest.config.ts
@@ -2,6 +2,11 @@ import { defineConfig } from "vitest/config";
 
 export default defineConfig({
   test: {
-    exclude: ["test/integration/**", "node_modules/**", "dist/**"],
+    exclude: [
+      "test/integration/**",
+      "node_modules/**",
+      "example-project/**",
+      "dist/**",
+    ],
   },
 });